Make `PatternColumn` part of the public API
